+++ Mods ChoGGi/Fix Tourism Update Colonist Issues/Code/Script.lua
@@ -5,13 +5,7 @@ if LuaRevision > 1001514 then
 end
 
 -- figure I just use the code from the hotfix
-function SavegameFixups.AddSatisfactionStat()
-  MapForEach(true, "Colonist", function(col)
-    col.stat_satisfaction = 0
-    col.log_satisfaction = {}
-  end)
-end
-function SavegameFixups.AddSatisfactionStat2()
+function SavegameFixups.AddSatisfactionStat1234567890()
   MapForEach(true, "Colonist", function(col)
     if not col.stat_satisfaction then
       col.stat_satisfaction = 0