+++ Mods ChoGGi/Unlock Breakthroughs/Code/Script.lua
@@ -28,11 +28,9 @@ end
 
 -- fired when Mod Options>Apply button is clicked
 function OnMsg.ApplyModOptions(id)
-	if id ~= CurrentModId then
-		return
+	if id == CurrentModId then
+		ModOptions()
 	end
-
-	ModOptions()
 end
 
 OnMsg.ModsReloaded = UpdateBreakthroughs