+++ Mods ChoGGi/Unlock Research Tech/metadata.lua
@@ -1,0 +1,24 @@
+return PlaceObj("ModDef", {
+	"title", "Unlock Research Tech",
+	"id", "ChoGGi_UnlockResearchTech",
+	"lua_revision", 1001551,
+	"steam_id", "2429688365",
+	"pops_any_uuid", "a2d712d9-57d2-47cd-bc9d-f09eeceb2193",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"TagOther", true,
+	"description", [[Unlock research tech in mod options.
+
+Other Mod Options:
+Tech Researched: Turn on to research instead of unlock tech.
+Always Apply Options: Unlock/Research techs whenever you load a game/start a new game.
+]],
+})